+++ scripts/cli/src/features/GenerateWebinyPkgCommand/index.ts
@@ -1,1 +1,2 @@
+export * from "./GenerateWebinyPkg.js";
 export * from "./GenerateWebinyPkgCommand.js";
